// chat-app/client/src/events/events.ts
export const enum Events {
  CONNECT = "connection",
  DISCONNECT = "disconnect",
  SEND_MESSAGE = "send_message",
  RECEIVE_MESSAGE = "receive_message",
  SET_NAME = "set_name",
  USER_JOINED = "user_joined",
  USER_LEFT = "user_left",
  USER_TYPING = "user_typing",
  USER_STOPPED_TYPING = "user_stopped_typing",
}